+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 edbee.lib:
 
+- add #56, Added move line up/down commands.
 - fix #58, Edbee crashes on Ctrl+Shift+Enter
+- Made ChangeGroup non-virtual. Sometimes you need a group of undo-changes in stead of a mergable one (issues with move-line command)
 - fix, Bug in Onig RegExp.cpp caused Ctrl+D to execute a regexp instead of a fixed string.
 - fix #54, Use of raw string literal causes issues when QT_NO_CAST_FROM_ASCII defined
 - fix #44, Commenting shortcut does not work if cursor is on the last character of the last line in the document
-- ref #45, Added ctrl-insert and shift-insert for copy paste
+- add #45, Added ctrl-insert and shift-insert for copy paste
 - fix #43, Added loadWithoutOpening and saveWithoutOpening to TextDocumentSerializer. For serializing without opening and closing an QIODevice (To enable the usage of QSaveFile).
 - fix #41, Added LUA comments to the hardcoded list of comments
 - fix #39, Margin-component bugfixes/improvements